+++ src/pages/Detail.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Detail = () => {
+  return (
+    <>
+      <h1>Detail</h1>
+    </>
+  );
+};
+
+export default Detail;
