+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as Home } from "./Home";
+export { default as Detail } from "./Detail";
